+++ CMakeLists.txt
@@ -538,7 +538,8 @@ if (NOT (CMAKE_BUILD_TYPE STREQUAL "Release") AND NOT DEFINED ENV{BODO_SKIP_CPP_
     "bodo/tests/test_external_sort.cpp"
     "bodo/tests/test_dict_builder.cpp"
     "bodo/tests/test_groupby_common.cpp"
-    "bodo/tests/test_groupby_and_window.cpp"
+    "bodo/tests/test_groupby.cpp"
+    "bodo/tests/test_window.cpp"
     "bodo/tests/test_json_col_reader.cpp"
     "bodo/tests/test_memory_budget.cpp"
     "bodo/tests/test_pinnable.cpp"
